pyproject.toml
[tool.poetry]
name = "dataminds-cassini-solution"
version = "0.1.0"
description = ""
authors = ["Panagiotis Papaemmanouil <papaemman.pan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
eodag = "^2.12.1"
geopandas = "^1.0.1"
rasterio = "^1.3.11"
matplotlib = "^3.9.2"
requests-oauthlib = "^2.0.0"
cdsapi = "^0.7.3"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
